[component] Created generic Notepad component
