+++ src/Components/Notepad.tsx
@@ -1,23 +1,12 @@
-import { getAllStudyLogs } from '@/lib/data';
+import { NotepadProps } from '@/lib/types';
 import Window from './Window';
 
-const getData = async () => {
-  const data = await getAllStudyLogs();
-  return data;
-};
-
-const Notepad = async () => {
-  const studyLogs = await getData();
+function Notepad(props: NotepadProps) {
   return (
     <Window English='Recently Studied' Japanese='最近の勉強'>
-      <div className='flex flex-col bg-slate-200 p-4 border-2 border-dark-gray'>
-        {studyLogs.map((log) => (
-          <span key={log.id}>• {log.title}</span>
-        ))}
-        <span>• | </span>
-      </div>
+      <div className='flex flex-col bg-slate-200 p-4 border-2 border-dark-gray'>{props.children}</div>
     </Window>
   );
-};
+}
 
 export default Notepad;
