+++ src/app/home/page.tsx
@@ -1,11 +1,11 @@
 import Graph from '@/Components/Graph';
-import Notepad from '@/Components/Notepad';
+import TodaysStudies from '@/Components/TodaysStudies';
 import Welcome from '@/Components/Welcome';
 
 export default function HomePage() {
   return (
     <main className='flex flex-col gap-4'>
-      <Notepad />
+      <TodaysStudies />
       <Graph />
       <Welcome />
     </main>
